requirements.txt
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0


# Nuevas dependencias para análisis de sentimiento
tweepy>=4.12.0
lunarcrush-python>=1.0.0
python-dateutil>=2.8.2
cachetools>=5.0.0
aiohttp>=3.8.1
